work with Jpa

+++ src/main/java/com/example/restful_api_product_spring_boot/service/ProductService.java
@@ -13,5 +13,5 @@ public interface ProductService {
     ProductResponse findProductById(Integer id);
 
     ProductResponse updateProduct(Integer id , UpdateProductRequest request);
-    boolean deleteProduct(Long id);
+    boolean deleteProduct(Integer id);
 }
